Separation of tests from each location
